check if auth enabled

+++ ci/tasks/hugo.sh
@@ -10,4 +10,8 @@ cd documentation-repo
 mv public ../compiled-site
 mv Staticfile ../compiled-site/
 mv manifest.yml ../compiled-site/
-cp ../password-site/Staticfile.auth ../compiled-site/
+
+if [ -f ../password-site/Staticfile.auth ]; then
+   echo "Password file found - copying!"
+   cp ../password-site/Staticfile.auth ../compiled-site/
+fi
